conUDS: Correct linux requirement and support non-linux builds in ota-agent
